Move `asUsd` and `validateMoney` extensions to the `MoneyExts` file.
