+++ client/src/main/kotlin/io/spine/examples/shareaware/client/StringExts.kt
@@ -27,25 +27,6 @@
 package io.spine.examples.shareaware.client
 
 import io.spine.examples.shareaware.wallet.Iban
-import io.spine.money.Currency
-import io.spine.money.Money
-
-/**
- * Returns a new `Money` object in USD currency using this `String` to construct it.
- *
- * This `String` must be written as a number with a decimal point.
- */
-public fun String.asUsd(): Money {
-    val parts = this.split('.')
-    val units = parts[0].toLong()
-    val nanos = if (parts.size == 2) parts[1].toInt() else 0
-    return Money
-        .newBuilder()
-        .setCurrency(Currency.USD)
-        .setUnits(units)
-        .setNanos(nanos)
-        .vBuild()
-}
 
 /**
  * Returns a new IBAN using this `String` as its value.
@@ -66,15 +47,6 @@ public fun String.validateIban(): Boolean {
     val ibanRegex =
         """[A-Z]{2}[0-9]{2}(?:[ ]?[0-9]{4}){4}(?!(?:[ ]?[0-9]){3})(?:[ ]?[0-9]{1,2})?""".toRegex()
     return !ibanRegex.containsMatchIn(this)
-}
-
-/**
- * Returns true if this `String` is written like a number with a decimal point,
- * and it can be converted to a `Money` object, false otherwise.
- */
-public fun String.validateMoney(): Boolean {
-    val decimalRegex = """^\d+(\.\d{1,2})?${'$'}""".toRegex()
-    return !decimalRegex.containsMatchIn(this)
 }
 
 /**
